+++ CHANGELOG.md
@@ -71,6 +71,10 @@ fn version_0_5() -> impl UiNode {
 * Fix panic on window move in Wayland.
 * Fix minimize command removing maximized state from restore.
 * Fix issue when parent widget's cursor can override child's cursor when the parent cursor var updates.
+* **Breaking** Remove the `cursor_img` property and window var.
+* **Breaking** The `cursor` property now has input type `CursorSource`.
+    - Note that the `CursorIcon` type converts to `CursorSource`.
+* Implement custom cursor images in the default view.
 
 # 0.4.0